Create 16S_QIIME_1
This first chunk of code combines forward and reverse reads and trims the primers from the reads. The trimmed-summary.qzv file produced, can be dropped into QIIME2view, to look at the interactive quality plots. The plot will help you to decide where to trim the forward and reverse reads.  Ideally you want to aim for a medium quality score >20 and an overlap of >20 bases.