Create 16S_QIIME2
This second chunk of code uses dada2 to denoise the reads, creates a feature table for the OTUs and generates a tree for phylogenetic diversity analyses.  The table.qvz file produced can then be dropped into qiime2view to decide the sampling depth for alpha and beta diversity analysis. The aim is to get the highest possible sampling depth without excluding too many samples.